[Add]更新

+++ Assets/Saruwatari/Script/IGauge.cs
@@ -4,7 +4,5 @@ using UnityEngine;
 
 public interface IGauge
 {
-   
-        void Gauge(int PlayerNumber);
-    
+        void Gauge(float PlayerNumber);   
 }
